Refactoring toolbar to same file
